Удаляет старые компоненты а переименовывает компоненты с суфиксом New на обычные
