+++ src/components/city-list/city-list.jsx
@@ -1,27 +1,32 @@
 import React from 'react';
-import {CITY_LIST_TYPE} from '../../types/types';
+import {connect} from 'react-redux';
+import {ActionCreator} from '../../store/action';
+import withActiveItem from '@hocs/with-active-item/with-active-item';
+import {CITY_LIST_TYPE} from '@types';
 
 const CityList = (props) => {
-  const {city, cities, onCityChange} = props;
+  const {cities, onCityChange, activeItem, onActiveItemChange} = props;
+  const activeCity = activeItem ? activeItem : cities[0];
 
   const handleCityClick = (evt) => {
     evt.preventDefault();
 
     const newCity = evt.target.innerText;
 
-    if (newCity !== city) {
+    if (newCity !== activeCity) {
       onCityChange(newCity);
+      onActiveItemChange(newCity);
     }
   };
 
-  const citiesList = cities.map((nextCity, i) => (
-    <li key={`${city}-${i}`} className="locations__item">
+  const citiesList = cities.map((city, i) => (
+    <li key={i} className="locations__item">
       <a
-        className={`locations__item-link tabs__item ${city === nextCity && `tabs__item--active`}`}
+        className={`locations__item-link tabs__item ${activeCity === city && `tabs__item--active`}`}
         href="#"
         onClick={handleCityClick}
       >
-        <span>{nextCity}</span>
+        <span>{city}</span>
       </a>
     </li>
   ));
@@ -35,7 +40,19 @@ const CityList = (props) => {
   );
 };
 
+
 CityList.propTypes = CITY_LIST_TYPE;
 
-export default CityList;
+const mapStateToProps = (state) => ({
+  city: state.city,
+  cities: state.cities,
+});
+const mapDispatchToProps = (dispath) => ({
+  onCityChange(city) {
+    dispath(ActionCreator.changeCity(city));
+    dispath(ActionCreator.getOffers(city));
+  },
+});
 
+export {CityList};
+export default connect(mapStateToProps, mapDispatchToProps)(withActiveItem(CityList));
