+++ src/components/main-page/main-page.jsx
@@ -1,20 +1,10 @@
 import React from 'react';
 import {connect} from "react-redux";
-
-// import OfferList from '../offer-list/offer-list';
-import OfferListNew from '../offer-list/offer-list-new';
-
-// import Map from '../map/map';
-import MapNew from '../map/map-new';
-
-// import CityList from '../city-list/city-list';
-import CityListNew from '../city-list/city-list-new';
-
-// import Sort from '../sort/sort';
-import SortNew from '../sort/sort-new';
-
+import OfferList from '../offer-list/offer-list';
+import Map from '../map/map';
+import CityList from '../city-list/city-list';
+import Sort from '../sort/sort';
 import Header from '../header/header';
-// import {getCityCenter} from '../../mocks/offers';
 import {MAIN_PAGE_TYPE} from '@types';
 import {OfferCardType, MapType} from '@const';
 import {FROM_SORTINGTYPE_TO_FUNC_MAP} from '@utils';
@@ -30,7 +20,7 @@ const MainPage = ({city, offers, sort}) => {
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
 
-          <CityListNew />
+          <CityList />
 
         </div>
         <div className="cities">
@@ -39,11 +29,11 @@ const MainPage = ({city, offers, sort}) => {
               <h2 className="visually-hidden">Places</h2>
               <b className="places__found">{offers.length} places to stay in {city}</b>
 
-              <SortNew />
+              <Sort />
 
               <div className="cities__places-list places__list tabs__content">
 
-                <OfferListNew
+                <OfferList
                   offers={offersSort}
                   type={OfferCardType.CITY_PLACE}
                 />
@@ -53,7 +43,7 @@ const MainPage = ({city, offers, sort}) => {
             <div className="cities__right-section">
               <section className="cities__map map">
 
-                <MapNew
+                <Map
                   key={city}
                   offers={offers}
                   city={city}
