+++ src/components/sort/sort.jsx
@@ -1,4 +1,6 @@
-import React from 'react';
+import React, {useState} from 'react';
+import {connect} from 'react-redux';
+import {ActionCreator} from '../../store/action';
 import {SORT_TYPE} from '../../types/types';
 import {SortingType} from '@const';
 
@@ -10,22 +12,30 @@ const sortValues = {
 };
 
 const Sort = (props) => {
-  const {sortActive, onSortChange} = props;
+  const {onSortChange} = props;
+  const [sort, setSort] = useState(SortingType.POPULAR);
+  const [isMenuOpen, toggleMenuOpen] = useState(true);
 
-  const onClick = (evt) => {
+  const handleSortChange = (evt) => {
     evt.preventDefault();
 
-    const newSortActive = evt.target.dataset.sort;
+    const newSort = evt.target.dataset.sort;
 
-    if (newSortActive !== sortActive) {
-      onSortChange(newSortActive);
+    if (newSort !== sort) {
+      setSort(newSort);
+      onSortChange(newSort);
     }
+  };
+
+  const handleMenuToggle = (evt) => {
+    evt.preventDefault();
+    toggleMenuOpen(!isMenuOpen);
   };
 
   const list = Object.entries(sortValues).map(([key, value]) => {
     return <li
       key={`${key}`}
-      className={`places__option ${key === sortActive && `places__option--active`}`}
+      className={`places__option ${key === sort && `places__option--active`}`}
       data-sort={key}
       tabIndex="0">
       {value}
@@ -35,13 +45,13 @@ const Sort = (props) => {
   return (
     <form className="places__sorting" action="#" method="get">
       <span className="places__sorting-caption">Sort by &nbsp;</span>
-      <span className="places__sorting-type" tabIndex="0">
-        {sortValues[sortActive]}
+      <span className="places__sorting-type" tabIndex="0" onClick={handleMenuToggle}>
+        {sortValues[sort]}
         <svg className="places__sorting-arrow" width="7" height="4">
           <use xlinkHref="#icon-arrow-select"></use>
         </svg>
       </span>
-      <ul onClick={onClick} className="places__options places__options--custom places__options--opened">
+      <ul onClick={handleSortChange} className={`places__options places__options--custom ${isMenuOpen && `places__options--opened`}`}>
         {list}
       </ul>
     </form>
@@ -50,5 +60,11 @@ const Sort = (props) => {
 
 Sort.propTypes = SORT_TYPE;
 
-export default Sort;
+const mapDispatchToProps = (dispath) => ({
+  onSortChange(sort) {
+    dispath(ActionCreator.changeSort(sort));
+  },
+});
 
+export {Sort};
+export default connect(null, mapDispatchToProps)(Sort);
