+++ src/store/action.js
@@ -1,12 +1,8 @@
 export const ActionType = {
   CHANGE_CITY: `CHANGE_CITY`,
   GET_OFFERS: `GET_OFFERS`,
-  // CHANGE_OFFER: `CHANGE_OFFER`,
-  // RESET_ACTIVE_OFFER_ID: `RESET_ACTIVE_OFFER_ID`,
+  CHANGE_OFFER: `CHANGE_OFFER`,
   CHANGE_SORT: `CHANGE_SORT`,
-
-  CHANGE_SORT_NEW: `CHANGE_SORT_NEW`,
-  CHANGE_OFFER_NEW: `CHANGE_OFFER_NEW`,
 };
 
 export const ActionCreator = {
@@ -18,14 +14,12 @@ export const ActionCreator = {
     type: ActionType.GET_OFFERS,
     payload: city,
   }),
-
-
-  changeSortNew: (sort) => ({
-    type: ActionType.CHANGE_SORT_NEW,
+  changeSort: (sort) => ({
+    type: ActionType.CHANGE_SORT,
     payload: sort,
   }),
-  changeOfferNew: (offer) => ({
-    type: ActionType.CHANGE_OFFER_NEW,
+  changeOffer: (offer) => ({
+    type: ActionType.CHANGE_OFFER,
     payload: offer,
   }),
 };
